MemSeg/requirements-forge.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
bokeh=0.12.11=py35h969371a_0
ca-certificates=2017.08.26=h94faf87_0
certifi=2017.7.27.1=py35hbab57cd_0
click=6.7=py35h10df73f_0
cloudpickle=0.5.2=py35hd830302_0
cycler=0.10.0=py35hcc71164_0
dask=0.16.0=py35hc436b91_0
dask-core=0.16.0=py35h0988329_0
decorator=4.1.2=py35hc7d1f2e_0
distributed=1.20.1=py35h3ced2a0_0
freetype=2.8=vc14h17c9bdf_0
heapdict=1.0.0=py35h3d1311d_0
icc_rt=2017.0.4=h97af966_0
icu=58.2=vc14hc45fdbb_0
imageio=2.2.0=py35hcd4b9a4_0
intel-openmp=2018.0.0=hcd89f80_7
jinja2=2.10=py35hdf652bb_0
jpeg=9b=vc14h4d7706e_1
libpng=1.6.32=vc14h5163883_3
libtiff=4.0.8=vc14h04e2a1e_10
locket=0.2.0=py35h0dfcdd0_1
markupsafe=1.0=py35hc253e08_1
matplotlib=2.1.0=py35hea29350_0
mkl=2018.0.0=h36b65af_4
msgpack-python=0.4.8=py35hdef45cb_0
networkx=1.11=py35h097edc8_0
numpy=1.14.0=py35h4a99626_1
olefile=0.44=py35h7ba86d0_0
openssl=1.0.2l=vc14hf4c37d5_5
pandas=0.21.0=py35h0510043_1
partd=0.3.8=py35h894d1e4_0
pillow=4.2.1=py35hd7da350_0
pip=9.0.1=py35h691316f_4
psutil=5.4.1=py35h5f28cdc_0
pyparsing=2.2.0=py35hcabcaab_1
pyqt=5.6.0=py35hd46907b_5
python=3.5.4=h1357f44_23
python-dateutil=2.6.1=py35h6b299a3_1
pytz=2017.2=py35h3d6be01_1
pywavelets=0.5.2=py35h7c47ace_0
pyyaml=3.12=py35h4bf9689_1
qt=5.6.2=vc14h6f8c307_12
scikit-image=0.13.1=py35ha2243b5_0
scipy=1.0.0=py35h75710e8_0
setuptools=36.5.0=py35h21a22e4_0
sip=4.18.1=py35h01cbaa7_2
six=1.11.0=py35hc1da2df_1
sortedcontainers=1.5.7=py35h9857911_0
sqlite=3.20.1=vc14h7ce8c62_1
tblib=1.3.2=py35hd2cf7e1_0
tifffile=0.13.5=py35_1
toolz=0.8.2=py35h7afdc4e_0
tornado=4.5.2=py35h03246f3_0
vc=14=h2379b0c_2
vs2015_runtime=14.0.25123=hd4c4e62_2
wheel=0.29.0=py35hdbcb6e6_1
wincertstore=0.2=py35hfebbdb8_0
yaml=0.1.7=hc54c509_2
zict=0.1.3=py35hf5542e0_0
zlib=1.2.11=vc14h1cdd9ab_1
